test: move budget expect into test file
