+++ packages/user-flow-cli-testing/src/index.ts
@@ -1,4 +1,3 @@
-export * from '../test/jest/expect.collect.budget';
 export * from './lib/constants';
 export * from './lib/types';
 export * from './lib/utils/cli-mode';
